Abstracting and Genericizing Controller, Part II
I am continuing work on abstracting and genericizing the controller.

I realized that I had not implemented a way for the journals to be posted to the General Ledger. To that end, I added two (2) more methods to the `JournalInterface`:

* `getGeneralLedgerEntry`, which creates an `EntryModel` object for the General Ledger.
* `isPosted`, which lets anyone creating a model for use with the `AbstractJournal` as the controller know they need to keep track of whether or not their journal model has been posted to the General Ledger.

I also added the `postToGeneralLedger` method to `AbstractJournal` to allow the journal controller to add entries to the General Ledger, if needed.

As you can see, work continues.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>