Abstracting and Genericizing Controller, Part III
Made some minor changes to the `AbstractJournal` and the `JournalInterface`.

In `AbstractJournal`, I moved all of the progress bar stuff into their own methods:

* `setUpProgressBar` sets up the progress bar when reading in the data from file.
* `advanceProgressBar` increases the progress bar value by one each time it is called.
* `setUpReverseProgressBar` sets up the progress bar when writing data to the file.
* `reduceProgressBar` reduces the progress bar value by one each time it is called.

I also renamed `createAndAddRecord` method to `loadRecord`, just to make things relate to the `JournalInterface`.

On the `JournalInterface`, I renamed `getRecordLine` method to `createRecord`, which better describes what is taking place.

Like I said, just minor modifications for now. I am just trying to find ways that I can make things easier on us in the future when we are maintaining the code base for the project and have users around the world.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>