Abstracting and Genericizing Controller, Part I
This is the first commit and push to re-create the `scarrick/FuelEntry` branch on the repository, without all of the documentation.

At this time, this branch will fail on the Travis CI tests, but @jkovalsy is working on updating his test causing the failure, so it should pass once he pushes the update to this branch.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>
Signed-off-by:Sean Carrick <sean@pekinsoft.com>