Abstracting and Genericizing Controller, Part II
## Cosmetic Changes

I have modified a lot of the controllers in the `com.pekinsoft.northwind.controller` package, cleaning up the file headers that were messed up a little in their formatting. But, this is just cosmetics.

## Fuel Purchase Dialog

On the `FuelPurchaseDialog`, I have added the ability to pay for the fuel purchase from, not only the Fuel Card account, but also from the Maintenance Reserve, Checking, and Savings accounts, in the event that there is no money on the Fuel Card, and the user is not under a load, but needs fuel. Since I have the system set to only allow cash advances when under an active load, I figured there should be a way for the user to purchase fuel, even if there is no money on their fuel card and they are not under a load where they could take a cash advance. Or, if they have already depleted the amount of cash advance they are allowed to take on a given load.

I, also, removed a bunch of unused imports from this class.

## Data Controllers

On all of my controllers, I have removed unnecessary casting that I had in place. I had put the casting in place because I did not yet have the full understanding of the generics I have used in the `AbstractJournal<T>` class that I created. I have gained some more understanding of generics since I implemented it and discovered that I did not need to cast the `ArrayList<T> records`, as they were already of the type I needed.

## String Utilities

Added the method `getMiddle(String source, String start, String end)` to the `com.pekinsoft.loadmaster.utils.StringUtils` class. Since we are parsing out account and ID numbers from strings in the `JComboBox`es throughout the project, I thought it would be a good idea to centralize that parsing functionality. Now, we do not need to make sure that we only use square brackets ([ and ]) around our numbers, as we can provide whatever character we use.

## String Utilities Test

I created a JUnit test class for testing `StringUtils` to verify that all of the methods in that class function as expected. I am having trouble running those tests right now, which is probably a configuration issue in my IDE. I will look into this and update status on next commit.

## Conclusion

This is everything that I have done thus far. I will continue working on this and make another commit later today.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>
Signed-off-by:Sean Carrick <sean@pekinsoft.com>