Restructuring Local Repository
Due to me being sToOpId, I need to restructure my local repository because of not switching back to the master branch prior to creating a new branch to modify existing code, or add new features. This is the only reason for this commit and push. I had started modifying the `Utils` class without creating a new branch. No changes are present, but I want to get rid of the commit notification.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>