Utils Class Update
Since we are attempting to build a professional, high-quality, integrated load tracking and accounting system for owner/operator truck drivers, I have updated the `com.pekinsoft.loadmaster.utils.Utils` class to help us achieve this end.

As we all know, each operating system has a standard for where on the system applications are supposed to store data, preferences, and log files. For this reason, I have added the following four (4) classes to the `Utils` class:

- `getApplicationFolderByOS`: checks the operating system name from the Java VM and returns a string of the path where that OS prefers applications to store their custom files.
- `getApplicationDataFolderByOS`: checks the operating system name from the Java VM and returns a string of the path where that OS prefers application to store their data files.
- `getApplicationSettingsLocationByOS`: checks the operating system name from the Java VM and returns a string of the path where that OS prefers applications to store their settings and configuration files.
- `getSystemLogLocationByOS`: checks the operating system name from the Java VM and returns a string of the path where that OS prefers applications to place their log files.

With this in mind, if we truly want a polished, professional application, we will need to update our code that sets up our folder structure and `AppLogger` class. Using these methods, we can place all of our Load Master™-generated files in the appropriate location for the system on which the user is running it.

@jkovalsy , if you approve of this, please just merge this branch to master. Otherwise, feel free to delete this branch and close the PR without merging. Thank you.

-SC
Signed-off-by:sean <sean@laptop.pekinsoft.com>
Signed-off-by:Sean Carrick <sean@pekinsoft.com>