Abstract Journal
I have created the `AbstractJournal` class for genericizing, as much as possible our, controller classes. Syntacticly, there are no errors showing, but logic and unit testing still needs to be created.

## Fuel Purchase Controller

I have updated the `FuelPurchaseCtl` class to extend the `AbstractJournal` class. Again, no syntactical errors, but untested and need logic and unit tests created. In fixing the syntax errors, I created the next class.

## Fuel Card Model

I created the `FuelCardModel` class on this branch because it was required by the `FuelPurchaseCtl` class. Because of this, I think that I will just create all of the remaining models and controllers on this branch. Once the testing is created and passed, the whole of the branch can be merged the master at the same time.

## Conclusion

I am calling this the first commit for the abstraction and accounting branch. I will continue working on this branch until there is one Model and one Controller for each journal in the accounting system.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>