Abstract Journal, Part II
## `com.pekinsoft.loadmaster.controller`

I have created all of the controller classes that will be required for the application. I have only implemented the constructor in each one so far, but will continue to work on these.

## `com.pekinsoft.loadmaster.model`

I have created all of the model classes that will be required for the application. I have only implemented the two `public static final` fields in these classes, but will implement each of them, then their related controller.

## Conclusion

Hopefully, I will have all of these classes fully implemented by the end of October, or first week in November. Once these classes are fully implemented, I will be able to start working on the accounting package, which will provide the methods for calculating the revenue and expenses per mile, in general, and for each Journal. Regardless, we will be to our first testing release once these classes are all implemented, and there are UI components that allow the user to use all of this stuff.

-SC
Signed-off-by:Sean Carrick <sean@pekinsoft.com>