Updated Project Website, Part II
I pulled the master branch to this branch after the upgrade of JUnit to 1.13 to fix the WhiteSource failure. I could not get the test to rerun, so I cleaned up the header of this file to give me a commit in order to re-push. Hopefully, it will pass all tests this time.

-SC
Signed-off-by:= <=>