Minor Change to Exit Functionality
I changed the logic of exiting to look at the size of an `EntryCtl` object in the `Starter` class, instead of using a project setting for the number of unprocessed batch transactions. 

With this change, adding entries to their respective journals will automatically create a General Ledger entry. However, the method for adding entries to journals will determine whether to post the transaction to the GL or to batch them. If the project setting is to use the Batch Posting Method, then the method that saves the transactions to their respective journals will create the GL entry in the `Starter.batch` `EntryCtl` object and the application `exit` methods in `LoadMaster` main window will check the size of that object to determine whether or not there are unprocessed transactions.

This will be much easier to maintain than the need of updating a counter in the application settings file.

-SC
Signed-off-by:Sean Carrick <sean at pekinsoft dot com>