Modified Utils.java
I have made the modifications to `Utils.java` by adding the following methods:

- `	getApplicationFolderByOS()`: This method makes sure that we always store our application files in the appropriate location on the host system.
- `getApplicationDataFolderByOS()`: This method makes sure that we always store our application's data files in the appropriate location on the host system.
- `getApplicationSettingsLocationByOS()`: This method makes sure that we always store our application's configuration files in the appropriate location on the host system.
- `getSystemLogLocationByOS()`: This method makes sure that we always store our application's log files in the appropriate location on the host system.

Why these changes?

Simply put, we want our application to follow common conventions when run on the various operating systems that we support. To this end, we don't want to clutter up the user's home folder with files and folders that should be stored in specific locations on the hard drive of the host operating system. These locations are set as follows:

| OS | Application Home | Application Data | Application Settings | Application Logs |
| -- | ---------------- | ---------------- | -------------------- | ---------------- |
| Windows | %UserHome%\AppData\LoadMaster\ | %UserHome%\AppData\LoadMaster\data\ | %UserHome%\AppData\LoadMaster.evt\ | %SystemRoot%\System32\Config\LoadMaster.evt
| Mac | {user.home}/Library/Application Data/LoadMaster/ | {user.home}/Library/Application Data/LoadMaster/data/ | {user.home}/Library/Preferences/LoadMaster.pref | /Library/Logs/LoadMaster/application.log |
| Linux | {user.home}/.loadmaster/ | {user.home}/.loadmaster/data/ | {user.home}/.loadmaster/loadmaster.conf | /var/log/loadmaster/application.log |
| Solaris | {user.home}/.loadmaster/ | {user.home}/.loadmaster/data/ | {user.home}/.loadmaster/loadmaster.conf | /var/log/loadmaster/application.log |

By calling one of these methods, the appropriate location will be returned, as the methods grab the OS name from the `System.getProperty("os.name")` call. Based off of the OS name returned, the methods return the appropriate location for storing files for the application.

This PR is in replacement of PRs #66, #68, #77 and #78, which had all gotten messed up with `doxygen` generated files in them. I have made sure that this commit only contains the one file, `com.pekinsoft.loadmaster.utils.Utils.java`.

-SC
Signed-off-by:Sean Carrick <sean at pekinsoft dot com>