Color and Code Improvements
I have modified the colors for error background and foreground, as well as for the tip label on the `Customers` form, `NewCustomerDlg`, and `NewBrokerDialog`.

Furthermore, I have cleaned up some logic in the `InputVerifier` classes, as well as added a general "catch-all" `InputVerifier` class called `DataPresentVerifier`. 

The `DataPresentVerifier` class provides a means of verifying that input has been made on a field that cannot be validated in any other way, such as street address, city, first name, last name, company name, etc. These types of fields may be required to have input, but unlike states, phone number, postal codes, and email addresses, these fields cannot be validated to a standard. Therefore, a user could simply tab off of a required field without placing any data into it. With the `DataPresentVerifier`, we are able to verify that data has been typed into the required field that has no format. Now, if the user tries to tab off of a required field without entering data, the background and foreground colors of the field are modified, as is the background of the tip label and the user is prevented from leaving that field. Once the user enters data into the field, even if only a single character, when they press tab (or click into another field), the focus will be allowed to change and the colors are reset to the standard colors.

One problem: the verifiers work as expected on every field, except the Zip Code field. For some reason, if the field is empty, the colors change to the error colors, but focus is allowed to shift. However, if there is invalid data in the field, the focus will not shift. I am having a little trouble with this one, so any help is greatly appreciated. I have tried overriding the `InputVerifier.shouldYieldFocus` method and returning the return value of the `verify` method, but that didn't work. This tells me that there is something broken in the logic of the `verify` method. I will keep looking at this until I, or someone else, figures it out.

-SC
Signed-off-by:Sean Carrick <sean at pekinsoft dot com>